Target-validated: 32-bit-only ISR math, official JieLi toolchain support
- Replace uint64_t arithmetic in Sequencer_Tick gate calc and audio
  clock/block-end paths with exact 32-bit equivalents; pi32v2 objects
  confirm no __udivdi3/__moddi3 or float helpers in seq/audio.
- Dockerfile: add python3 (cc wrapper needs it), pin toolchain
  provenance (pkgman linux-toolchain).
- Makefile check-no-malloc excludes gitignored build/ scratch space.

+++ sequencer.c
@@ -236,7 +236,11 @@ void Sequencer_Tick(uint32_t current_bpm)
             } else if (gate > 100u) {
                 gate = 100u;
             }
-            uint32_t gate_us = (uint32_t)(((uint64_t)seq_step_dur * gate) / 100u);
+            /* Gate duration. 32-bit only: step_dur <= 500000 (30 BPM),
+             * gate <= 100, so the product (<= 5e7) cannot overflow.
+             * (A 64-bit multiply here would drag in __udivdi3 soft-divide
+             * on pi32v2, which has no 64-bit divider.) */
+            uint32_t gate_us = (seq_step_dur * (uint32_t)gate) / 100u;
             if (gate_us < 1000u) {
                 gate_us = 1000u; /* minimum 1 ms so offs never collapse onto ons */
             }
